feat: 明细表行头单元格支持拖拽
